{{ cookiecutter.project_name }}/composition_repository/lotus/requirements.txt
django-blog-lotus[breadcrumbs]==0.5.2
